Merge with upstream

+++ autobuild.sh
@@ -3,7 +3,7 @@
 set -e
 
 rootdir=$(readlink -f $(dirname $0))
-source "$rootdir/scripts/autotest_common.sh"
+source "$rootdir/test/common/autotest_common.sh"
 
 out=$PWD
 
@@ -13,6 +13,13 @@ cd $rootdir
 
 date -u
 git describe --tags
+
+# Print some test system info out for the log
+echo "** START ** Info for Hostname: $HOSTNAME"
+uname -a
+$MAKE cc_version
+$MAKE cxx_version
+echo "** END ** Info for Hostname: $HOSTNAME"
 
 timing_enter autobuild
 
@@ -24,20 +31,31 @@ if [ $SPDK_RUN_CHECK_FORMAT -eq 1 ]; then
 fi
 timing_exit check_format
 
-timing_enter build_kmod
-if [ $SPDK_BUILD_IOAT_KMOD -eq 1 ]; then
-	./scripts/build_kmod.sh build
-fi
-timing_exit build_kmod
-
 scanbuild=''
+make_timing_label='make'
 if [ $SPDK_RUN_SCANBUILD -eq 1 ] && hash scan-build; then
 	scanbuild="scan-build -o $out/scan-build-tmp --status-bugs"
+	make_timing_label='scanbuild_make'
+	report_test_completion "scanbuild"
+
 fi
+
+if [ $SPDK_RUN_VALGRIND -eq 1 ]; then
+	report_test_completion "valgrind"
+fi
+
+if [ $SPDK_RUN_ASAN -eq 1 ]; then
+	report_test_completion "asan"
+fi
+
+if [ $SPDK_RUN_UBSAN -eq 1 ]; then
+	report_test_completion "ubsan"
+fi
+
 echo $scanbuild
 $MAKE $MAKEFLAGS clean
 
-timing_enter scanbuild_make
+timing_enter "$make_timing_label"
 fail=0
 time $scanbuild $MAKE $MAKEFLAGS || fail=1
 if [ $fail -eq 1 ]; then
@@ -51,18 +69,21 @@ if [ $fail -eq 1 ]; then
 else
 	rm -rf $out/scan-build-tmp
 fi
-timing_exit scanbuild_make
+timing_exit "$make_timing_label"
 
 # Check for generated files that are not listed in .gitignore
-if [ `git status --porcelain | wc -l` -ne 0 ]; then
+timing_enter generated_files_check
+if [ `git status --porcelain --ignore-submodules | wc -l` -ne 0 ]; then
 	echo "Generated files missing from .gitignore:"
 	git status --porcelain
 	exit 1
 fi
+timing_exit generated_files_check
 
 # Check that header file dependencies are working correctly by
 #  capturing a binary's stat data before and after touching a
 #  header file and re-making.
+timing_enter dependency_check
 STAT1=`stat examples/nvme/identify/identify`
 sleep 1
 touch lib/nvme/nvme_internal.h
@@ -73,18 +94,36 @@ if [ "$STAT1" == "$STAT2" ]; then
 	echo "Header dependency check failed"
 	exit 1
 fi
+timing_exit dependency_check
 
+# Test 'make install'
+timing_enter make_install
+rm -rf /tmp/spdk
+mkdir /tmp/spdk
+$MAKE $MAKEFLAGS install DESTDIR=/tmp/spdk prefix=/usr
+ls -lR /tmp/spdk
+rm -rf /tmp/spdk
+timing_exit make_install
 
 timing_enter doxygen
 if [ $SPDK_BUILD_DOC -eq 1 ] && hash doxygen; then
-	(cd "$rootdir"/doc; $MAKE $MAKEFLAGS) &> "$out"/doxygen.log
-	if hash pdflatex; then
-		(cd "$rootdir"/doc/output/latex && $MAKE $MAKEFLAGS) &>> "$out"/doxygen.log
+	$MAKE -C "$rootdir"/doc --no-print-directory $MAKEFLAGS &> "$out"/doxygen.log
+	if [ -s "$out"/doxygen.log ]; then
+		cat "$out"/doxygen.log
+		echo "Doxygen errors found!"
+		exit 1
+	fi
+	if hash pdflatex 2>/dev/null; then
+		$MAKE -C "$rootdir"/doc/output/latex --no-print-directory $MAKEFLAGS &>> "$out"/doxygen.log
 	fi
 	mkdir -p "$out"/doc
 	mv "$rootdir"/doc/output/html "$out"/doc
 	if [ -f "$rootdir"/doc/output/latex/refman.pdf ]; then
 		mv "$rootdir"/doc/output/latex/refman.pdf "$out"/doc/spdk.pdf
+	fi
+	$MAKE -C "$rootdir"/doc --no-print-directory $MAKEFLAGS clean &>> "$out"/doxygen.log
+	if [ -s "$out"/doxygen.log ]; then
+		rm "$out"/doxygen.log
 	fi
 	rm -rf "$rootdir"/doc/output
 fi
